Modularized LandingPage
